FlyCam
Found this script which seems to work, though sometimes it spirals off out of control. Just one-lined most of the variables that were the same, but have not altered its functionality.
To use it just put it on the playerchar and uncheck the current player controller so that this one is being used. Uncheck it and check the other one to use that again.