Closer to Prop Switching
Main problem is that I don't know how to instantiate a new camera at a particular location. So perhaps every prop must have its own manually placed inactive camera that is activated when you switch to the copy of it. Other problem is I can't seem to reach the ThirdPerson component by name in order to be able to enable/disable it through script.