Warehouse continued
Stuck trying to paint materials. Also trying to make a side room at top of stairs but did not find a way to delete a small part of wall to make a door. I managed it before using the "Subdivide Faces" which worked on the window, but when trying on the side walls it does not divide into squares.