Shrinking Collider attempt
Made roundTime 20 in float value, and it shrinks till roundTime goes to 10 seconds, i.e. if roundTime > 10, shrink. Obviously the problem with that is that it is dependent on roundTime - and if players can customize roundTime then if they set it high enough the sphere will have already shrunk to practically nothing before the last 10 seconds of round anyway. Ah well, I think it can be good to demonstrate tomorrow.