2D graphics above zone doors
Trying to go for an "LCD" look. What I hope is that it will make it clear to test participants in week 9 and anyone else which area is what. 

Wasted some time looking for how to make texture act like a "scrolling marquee" (web design term) on a plane. Looks too complicated to waste more time on.

Maybe I will remove these later if I can't find a way to switch a green texture to a red one when the zone shuts off.