Retail (2)
First shop - ball item, since the only other place this is used is in the Park which is the zone furthest away from it. Whereas the items used in Residential should be closer to the other of Warehouse so that the player does not immediately see the same items upon entrance all over again. 

Pondered how to make this shop more interesting and so playing with idea of a video texture to "advertise" ball object inside shop. Problem is, .avi is the only format I can record in which Unity will accept. And so even a few seconds footage is already 20MB... tomorrow morning I will look more at the plausibility of a video texture on the screen in the shop.