Retail (3)
Did get video to play on screen object. Going to see how this will affect performance and then try some more. Retail can have several videos playing around it to represent different products. 

When I recorded, .avi was the only format I could record in that Unity would support, which still didn't work. So I used online-convert.com to make it an .mp4 which works fine and is like 1/4 of the size of the .avi...