Colliders
Changed most things to mesh collider in this version because of ridiculous bullet decals on non-cube objects. Also set a heap of 2D graphics to Crunch manually - even though there seems to be some sort of default setting since I last did this which set them all to "Compressed"??? Just don't update to this one if either of those things seems like a wrong move.