All warehouse props should work
Fridge model seems to still be part underground. 

Wet floor sign's shape means that none of the colliders are shaped to deal with it properly. So using Mesh Collider with convex - but if shooting at the sign from the side, decals will appear in mid-air. Box Collider would be useless as there is no way to rotate it to be diagonal.