All props playable???
Don't think I'm missing anything except a few that make no sense. 

Ball - Problem with this one is that if it were a playable prop it would seem wrong for it not to roll yet still move by rolling when shot or collided with. 

Flower Pot - The flower graphic faces the camera at all times which would look bizarre in third-person.

Obviously cups, cans, radio, etc. aren't playable. As I'm not sure how the health modifier is set to props they're all 0 or 1. So those tiny props would have a large advantage at the moment.